surveyRdbCreationJob: refactored folder structure
