+++ server/modules/surveyRdb/service/surveyRdbCreationJob/surveyRdbResultTablesCreationJob.js
@@ -14,10 +14,10 @@ export default class SurveyRdbResultTablesCreationJob extends Job {
 
     await SurveyRdbManager.createResultNodeTable(surveyId, tx)
 
-    const resultStepViewsByEntityUuid = await SurveyRdbManager.generateResultViews(surveyId, tx)
+    const resultStepViewsByEntityUuid = await SurveyRdbManager.getResultStepViews(surveyId, tx)
     const resultStepViews = R.pipe(R.values, R.flatten)(resultStepViewsByEntityUuid)
     await Promise.all(
-      resultStepViews.map(resultStepView => SurveyRdbManager.createResultStepView(surveyId, resultStepView, tx)),
+      resultStepViews.map((resultStepView) => SurveyRdbManager.createResultStepView(surveyId, resultStepView, tx))
     )
   }
 }
